+++ lib/src/data/data.dart
@@ -4,3 +4,4 @@ export 'indicator_state.dart';
 export 'indicator_state_change.dart';
 export 'indicator_trigger_edge.dart';
 export 'indicator_trigger_mode.dart';
+export 'refresh_indicator_durations.dart';