+++ lib/src/widgets/custom_material_indicator.dart
@@ -2,13 +2,23 @@ import 'package:custom_refresh_indicator/custom_refresh_indicator.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
+/// A function type that builds a material-style indicator widget.
+///
+/// This builder takes the current [BuildContext] and an [IndicatorController]
+/// to construct the widget.
 typedef MaterialIndicatorBuilder = Widget Function(
   BuildContext context,
   IndicatorController controller,
 );
 
 class CustomMaterialIndicator extends StatelessWidget {
+  /// {@macro custom_refresh_indicator.child}
   final Widget child;
+
+  /// {@macro custom_refresh_indicator.auto_rebuild}
+  final bool autoRebuild;
+
+  /// {@macro custom_refresh_indicator.on_refresh}
   final AsyncCallback onRefresh;
 
   /// The distance from the child's top or bottom [edgeOffset] where
@@ -45,17 +55,46 @@ class CustomMaterialIndicator extends StatelessWidget {
   /// Builds the content for the indicator container
   final MaterialIndicatorBuilder indicatorBuilder;
 
-  /// Builds the scrollable.
+  /// A builder that constructs a scrollable widget, typically used for a list.
+  ///
+  /// This builder is responsible for building the scrollable widget ([child])
+  /// that can be animated during loading or other state changes.
   final IndicatorBuilder scrollableBuilder;
 
   /// When set to *true*, the indicator will rotate
   /// in the [IndicatorState.loading] state.
   final bool withRotation;
 
+  /// {@macro custom_refresh_indicator.notification_predicate}
+  final ScrollNotificationPredicate notificationPredicate;
+
   /// The content will be clipped (or not) according to this option.
   ///
   /// Defaults to [Clip.none].
   final Clip clipBehavior;
+
+  /// {@macro custom_refresh_indicator.indicator_trigger}
+  ///
+  /// {@macro custom_refresh_indicator.trigger}
+  final IndicatorTrigger trigger;
+
+  /// {@macro custom_refresh_indicator.trigger_mode}
+  final IndicatorTriggerMode triggerMode;
+
+  /// {@macro custom_refresh_indicator.controller}
+  final IndicatorController? controller;
+
+  /// {@macro custom_refresh_indicator.durations}
+  final RefreshIndicatorDurations durations;
+
+  /// {@macro custom_refresh_indicator.on_state_changed}
+  final OnStateChanged? onStateChanged;
+
+  /// Whether to display leading scroll indicator
+  final bool leadingScrollIndicatorVisible;
+
+  /// Whether to display trailing scroll indicator
+  final bool trailingScrollIndicatorVisible;
 
   const CustomMaterialIndicator({
     Key? key,
@@ -63,12 +102,21 @@ class CustomMaterialIndicator extends StatelessWidget {
     required this.onRefresh,
     required this.indicatorBuilder,
     this.scrollableBuilder = _defaultBuilder,
+    this.notificationPredicate = CustomRefreshIndicator.defaultScrollNotificationPredicate,
     this.backgroundColor,
     this.displacement = 40.0,
     this.edgeOffset = 0.0,
     this.withRotation = true,
     this.elevation = 2.0,
     this.clipBehavior = Clip.none,
+    this.autoRebuild = true,
+    this.trigger = IndicatorTrigger.leadingEdge,
+    this.triggerMode = IndicatorTriggerMode.onEdge,
+    this.controller,
+    this.durations = const RefreshIndicatorDurations(),
+    this.onStateChanged,
+    this.leadingScrollIndicatorVisible = false,
+    this.trailingScrollIndicatorVisible = true,
   }) : super(key: key);
 
   static Widget _defaultBuilder(BuildContext context, Widget child, IndicatorController controller) => child;
@@ -77,8 +125,15 @@ class CustomMaterialIndicator extends StatelessWidget {
   Widget build(BuildContext context) {
     return CustomRefreshIndicator(
       autoRebuild: false,
-      leadingScrollIndicatorVisible: false,
+      notificationPredicate: notificationPredicate,
       onRefresh: onRefresh,
+      trigger: trigger,
+      triggerMode: triggerMode,
+      controller: controller,
+      durations: durations,
+      onStateChanged: onStateChanged,
+      trailingScrollIndicatorVisible: trailingScrollIndicatorVisible,
+      leadingScrollIndicatorVisible: leadingScrollIndicatorVisible,
       builder: (context, child, controller) {
         final Color backgroundColor = this.backgroundColor ??
             ProgressIndicatorTheme.of(context).refreshBackgroundColor ??
@@ -104,7 +159,12 @@ class CustomMaterialIndicator extends StatelessWidget {
                     elevation: elevation,
                     child: _InfiniteRotation(
                       running: withRotation && controller.isLoading,
-                      child: indicatorBuilder(context, controller),
+                      child: autoRebuild
+                          ? AnimatedBuilder(
+                              animation: controller,
+                              builder: (context, _) => indicatorBuilder(context, controller),
+                            )
+                          : indicatorBuilder(context, controller),
                     ),
                   ),
                 ),
@@ -166,34 +226,41 @@ class _PositionedIndicatorContainer extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    if (controller.side.isNone) return const SizedBox.shrink();
+    final side = controller.side;
+    if (side.isNone) return const SizedBox.shrink();
 
-    final isVerticalAxis = controller.side.isTop || controller.side.isBottom;
-    final isHorizontalAxis = controller.side.isLeft || controller.side.isRight;
+    final isVerticalAxis = side.isTop || side.isBottom;
+    final isHorizontalAxis = side.isLeft || side.isRight;
 
     final AlignmentDirectional alignment = isVerticalAxis
-        ? AlignmentDirectional(-1.0, controller.side.isTop ? 1.0 : -1.0)
-        : AlignmentDirectional(controller.side.isLeft ? 1.0 : -1.0, -1.0);
+        ? AlignmentDirectional(-1.0, side.isTop ? 1.0 : -1.0)
+        : AlignmentDirectional(side.isLeft ? 1.0 : -1.0, -1.0);
+
+    final endOffset = isVerticalAxis ? Offset(0.0, side.isTop ? 1.0 : -1.0) : Offset(side.isLeft ? 1.0 : -1.0, 0.0);
+
+    final animation = controller.isFinalizing
+        ? AlwaysStoppedAnimation(endOffset)
+        : Tween(begin: const Offset(0.0, 0.0), end: endOffset).animate(controller);
 
     return Positioned(
       top: isHorizontalAxis
           ? 0
-          : controller.side.isTop
+          : side.isTop
               ? edgeOffset
               : null,
       bottom: isHorizontalAxis
           ? 0
-          : controller.side.isBottom
+          : side.isBottom
               ? edgeOffset
               : null,
       left: isVerticalAxis
           ? 0
-          : controller.side.isLeft
+          : side.isLeft
               ? edgeOffset
               : null,
       right: isVerticalAxis
           ? 0
-          : controller.side.isRight
+          : side.isRight
               ? edgeOffset
               : null,
       child: Align(
@@ -201,13 +268,11 @@ class _PositionedIndicatorContainer extends StatelessWidget {
         heightFactor: isVerticalAxis ? 0.0 : null,
         widthFactor: isHorizontalAxis ? 0.0 : null,
         child: SlideTransition(
-          position: controller.isFinalizing
-              ? const AlwaysStoppedAnimation(Offset(0.0, 1.0))
-              : Tween(begin: const Offset(0.0, 0.0), end: const Offset(0.0, 1.0)).animate(controller),
+          position: animation,
           child: Padding(
-            padding: _getEdgeInsets(controller.side),
+            padding: _getEdgeInsets(side),
             child: Align(
-              alignment: _getAlignement(controller.side),
+              alignment: _getAlignement(side),
               child: child,
             ),
           ),